fastapi_server/requirements.txt
fastapi
uvicorn
pillow
pandas
pyyaml
pytest
httpx
python-multipart

transformers

numpy==1.24.4